+++ Dominio/Genero.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Genero
     {
-        public int ID { get; set; }
+        public byte ID { get; set; }
         public string Nombre { get; set; }
         public bool Estado { get; set; }
     }
